+++ isa-project/src/main/java/isaproject/service/impl/FishingCourseServiceImpl.java
@@ -316,13 +316,13 @@ public class FishingCourseServiceImpl implements FishingCourseService {
 	@Override
 	public IncomeDTO getFishingCourseIncomeYearly(DateTimeSpan duration, long id) {
 		long yearCount = duration.getYears() + 1;
-		int[] incomeSum = new int[(int) yearCount];
+		Double[] incomeSum = new Double[(int) yearCount];
 		Income income = new Income();
 		FishingCourse fishingCourse = courseRepository.findById(id).get();
 		Set<FishingReservation> reservations = fishingCourse.getFishingReservation();
 		for (FishingReservation reservation : reservations) {
 			for (int i = 1; i <= yearCount; i++) {
-				incomeSum = statisticsService.yearlyIncome(reservation.getDuration(), reservation.getPrice().intValue(), i,
+				incomeSum = statisticsService.yearlyIncome(reservation.getDuration(), reservation.getPrice(), i,
 						incomeSum, duration.getStartDate().getYear());
 			}
 
@@ -334,7 +334,7 @@ public class FishingCourseServiceImpl implements FishingCourseService {
 	@Override
 	public IncomeDTO getFishingCourseIncomeMonthly(DateTimeSpan duration, long id) {
 		long yearCount = duration.getYears() + 1;
-		int[][] incomeSum = new int[(int) yearCount][12];
+		Double[][] incomeSum = new Double[(int) yearCount][12];
 		Income income = new Income();
 		FishingCourse fishingCourse = courseRepository.findById(id).get();
 		Set<FishingReservation> reservations = fishingCourse.getFishingReservation();
@@ -342,7 +342,7 @@ public class FishingCourseServiceImpl implements FishingCourseService {
 			for (int i = 1; i <= yearCount; i++) {
 				for (int j = 1; j <= 12; j++) {
 					if (duration.isBetween(reservation.getDuration().getEndDate())) {
-						incomeSum = statisticsService.monthlyIncome(reservation.getDuration(),reservation.getPrice().intValue(),
+						incomeSum = statisticsService.monthlyIncome(reservation.getDuration(),reservation.getPrice(),
 								i, j, incomeSum, duration.getStartDate().getYear());
 					}
 				}
@@ -356,7 +356,7 @@ public class FishingCourseServiceImpl implements FishingCourseService {
 	@Override
 	public IncomeDTO getFishingCourseIncomeDaily(DateTimeSpan duration, long id) {
 		long yearCount = duration.getYears() + 1;
-		int[][][] incomeSum = new int[(int) yearCount][12][31];
+		Double[][][] incomeSum = new Double[(int) yearCount][12][31];
 		Income income = new Income();
 		FishingCourse fishingCourse = courseRepository.findById(id).get();
 		Set<FishingReservation> reservations = fishingCourse.getFishingReservation();
@@ -365,7 +365,7 @@ public class FishingCourseServiceImpl implements FishingCourseService {
 				for (int j = 1; j <= 12; j++) {
 					for (int k = 1; k <= 31; k++) {
 						if (duration.isBetween(reservation.getDuration().getEndDate())) {
-							incomeSum = statisticsService.dailyIncome(reservation.getDuration(), reservation.getPrice().intValue(),
+							incomeSum = statisticsService.dailyIncome(reservation.getDuration(), reservation.getPrice(),
 									i, j, k, incomeSum, duration.getStartDate().getYear());
 						}
 					}
